+++ src/types/types.cpp
@@ -249,7 +249,7 @@ struct FramesList
 // this list expires every time when program continues execution.
 using KnownFrames = Singleton<FramesList>;
 
-} // namespace
+} // unnamed namespace
 
 FrameId::FrameId(ThreadId thread, FrameLevel level)
     : m_id(static_cast<ScalarType>(KnownFrames::instance().get()->emplace(thread, level).first->first))